hard changes: first step of many steps
