+++ STRUCTURE.md
@@ -1,0 +1,28 @@
+
+
+# Structure
+
+
+## ex.py
+    
+- ex.py: main class for exercise
+- exsicua.py:
+- exlatex.py: 
+
+
+## megbook.py
+
+* repository of exercices
+* all markups at same time
+* SMC, command-line linux: a project can has only one megbook with all exercises.
+* user could not see explicit where the exercises are being stored.
+* the database can be a server database or sqlite
+
+
+
+## xporter.py
+
+* xlatex.py
+* xsiacua.py
+* xmoodle.py
+
